requirements.txt
vectorbt
requests  # Keep for now, though we won't need it after switching
pandas
python-dotenv  # If not already added
openalgo
anywidget
ta-lib
numpy
kaleido